Renamed Levi's alternate interface PnP problem solver to be the default. Added documentation to match. Adjusted use cases here and there.
