+++ rct_optimizations/CMakeLists.txt
@@ -39,7 +39,6 @@ add_library(${PROJECT_NAME}
   src/${PROJECT_NAME}/camera_intrinsic.cpp
   src/${PROJECT_NAME}/pnp.cpp
   src/${PROJECT_NAME}/multi_camera_pnp.cpp
-  src/${PROJECT_NAME}/multi_static_camera_pnp.cpp
 )
 
 add_dependencies(${PROJECT_NAME} ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
