+++ rct_optimizations/include/rct_optimizations/experimental/multi_camera_pnp.h
@@ -1,41 +1,72 @@
+/*
+ * Computes the pose of a "target" seen in one or more intrinsically and extrinsically calibrated
+ * cameras. The pose of these cameras must be specified in a common "base" frame, and the guess
+ * at the target must be specified in this same base frame. The answer will also be in this common
+ * base frame.
+ *
+ * Note that this function is logically equivalent to the solver available in pnp.h assuming that
+ * you specified ONE camera with an identity transform.
+ *
+ * author: Levi Armstrong
+ */
+
 #ifndef RCT_MULTI_CAMERA_PNP_H
 #define RCT_MULTI_CAMERA_PNP_H
 
-#include <rct_optimizations/types.h>
+#include "rct_optimizations/types.h"
 
 namespace rct_optimizations
 {
 
 struct MultiCameraPnPProblem
 {
-  // For each camera, define its intrinsics
+  /** @brief The basic camera intrinsic propeties: fx, fy, cx, cy used to reproject points;
+      one for each camera. Should have same length as @e base_to_camera. */
   std::vector<CameraIntrinsics> intr;
-  // For each camera, define the correspondences it saw
-  std::vector<CorrespondenceSet> correspondences;
 
-  // If you have more than 1 camera, this vector specifies the transform between the optical
-  // frame of the FIRST camera in the above image set and each subsequent camera.
-  // So camera_transforms[0] = camera 0 to camera 1
-  //    camera_ttransforms[1] = camera 0 to camera 2
-  // etc...
-  std::vector<Eigen::Affine3d> camera_transforms;
+  /** @brief The "base frame" to "camera frame" transform; one for each camera. Should have same
+      length as @e intr. */
+  std::vector<Eigen::Affine3d> base_to_camera;
 
-  // The transform from the FIRST camera to the target
-  Eigen::Affine3d camera_to_target_guess;
+  /** @brief A sequence of observation sets corresponding to the image locations.
+   * Each observation set consists of a set of correspodences: a 3D position (e.g. a dot) in "target
+   * frame" to the image location it was detected at (2D). The outer-most vector is for each camera,
+   * the inner vector is the images valid for that camera. */
+  std::vector<CorrespondenceSet> image_observations;
+
+  /** @brief Your best guess for transforms, "base to target", for a given observation set taken.*/
+  Eigen::Affine3d base_to_target_guess;
 };
 
 struct MultiCameraPnPResult
 {
+  /**
+   * @brief Whether the underlying solver converged. If this is false, your calibration did not go well.
+   * If this is true, your calibration MAY have gone well.
+   */
   bool converged;
+
+  /**
+   * @brief The initial reprojection error (in pixels) per residual based on your input guesses.
+   */
   double initial_cost_per_obs;
+
+  /**
+   * @brief The final reprojection error (in pixels) per residual after optimization. Note that each circle
+   * has two residuals: a U and V error in the image. So a value of 1.2 means that each circle was described
+   * to within 1.2 pixels in X and 1.2 pixels in Y.
+   *
+   * A low value here is encouraging if you had a diversity of images. If you took few images, you can get
+   * a low score without getting a calibration that describes your workcell.
+   */
   double final_cost_per_obs;
 
-  // The transform from the FIRST camera to the target
-  Eigen::Affine3d camera_to_target;
+  /** @brief The final location of the target. */
+  Eigen::Affine3d base_to_target;
 };
 
 MultiCameraPnPResult optimize(const MultiCameraPnPProblem& params);
 
 }
 
-#endif // RCT_MUTLI_CAMERA_PNP_H
+#endif // RCT_MULTI_STATIC_CAMERA_PNP_H
